+++ test/demos.ts
@@ -1,7 +1,7 @@
 import {mount, testDemos, unmount, wait, nextFrame} from './utils';
 
 const req = require.context('~/components/', true, /^((?!(affix|code)).)*\/demos\/.*index\.ts$/);
-// const req = require.context('~/components/', true, /upload\/demos\/.*index\.ts$/);
+// const req = require.context('~/components/', true, /button\/demos\/.*index\.ts$/);
 // const vueReq = require.context('~/components/', true, /^((?!(affix|code)).)*\/demos\/.*index\.vue$/);
 
 describe('Demos', () => {